lecta/requirements.txt
python-dotenv==1.0.1
django==5.0.4
djangorestframework==3.15.1
django-filter==23.5
psycopg2==2.9.9
django-cors-headers==3.14.0
drf_spectacular==0.26.1